[Frontend] Move notes/entities to notes folder
