+++ frontend/src/notes/entities.ts
