+++ frontend/src/notes/index.ts
@@ -1,4 +1,5 @@
 export * from './clients';
 export * from './interactors';
+export * from './entities';
 export * from './stores';
 export * from './ui';
