+++ frontend/src/notes/interactors/EditNoteInteractor.test.ts
@@ -1,7 +1,7 @@
 import { newEditNoteState, EditNoteInteractor } from './EditNoteInteractor';
 import { uuid } from '../../utils';
 import { InMemoryNoteClient } from '../clients';
-import { Note } from './entities';
+import { Note } from '../entities';
 
 describe('newEditNoteState', () => {
   it('returns an empty state', () => {
