+++ frontend/src/notes/interactors/index.ts
@@ -1,4 +1,3 @@
-export * from './entities';
 export * from './EditNoteInteractor';
 export * from './ListNoteInteractor';
 export * from './RemoveNoteInteractor';
